% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_gap_filll_trailing_NAs.R
\name{add_daily_cycle}
\alias{add_daily_cycle}
\title{extracts the daily cycle of the specified variable and joins it on the data,
as well as the deviation}
\usage{
add_daily_cycle(ts_dat, variable_name = "no2")
}
\arguments{
\item{ts_dat}{time series data.frame with a date_time column POSIXct}

\item{variable_name}{column_name of the variable for which the cycle should by calculated}
}
\description{
extracts the daily cycle of the specified variable and joins it on the data,
as well as the deviation
}
