% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{complement_times}
\alias{complement_times}
\title{complements a data frame with all hourly values in the respective
time frame of the date_time column}
\usage{
complement_times(ts_dat, fill_until = NULL)
}
\arguments{
\item{ts_dat}{time series data.frame, needs a date_time column in POSIX.ct format}

\item{fill_until}{time point to stop filling, if NULL the max of ts_dat is used.}
}
\description{
complements a data frame with all hourly values in the respective
time frame of the date_time column
}
