% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_gap_filll_trailing_NAs.R
\name{do_gap_filling_element}
\alias{do_gap_filling_element}
\title{do the gap filling to one element of an ordered list, each element with trailing NAs at the end
the function applies the gap filling to one element of the list}
\usage{
do_gap_filling_element(i, ts_dat_list, variable_name, daily_only = FALSE)
}
\arguments{
\item{i}{element of the list}

\item{ts_dat_list}{ordered list of trailing NA blocks}

\item{variable_name}{variable_name for gap filling}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
do the gap filling to one element of an ordered list, each element with trailing NAs at the end
the function applies the gap filling to one element of the list
}
