% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_fill_constant_variables.R
\name{fill_constant_variables}
\alias{fill_constant_variables}
\title{fill the constant variables (i.e. identifiers etc.)
by default all except variables defined in exclude_variables and date_time}
\usage{
fill_constant_variables(ts_dat, exclude_variables)
}
\arguments{
\item{ts_dat}{time series data frame of one location}

\item{exclude_variables}{which are excluded from filling (i. e. not constant)}
}
\description{
fill the constant variables (i.e. identifiers etc.)
by default all except variables defined in exclude_variables and date_time
}
