% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_gap_filll_trailing_NAs.R
\name{fill_large_gaps}
\alias{fill_large_gaps}
\title{fills gaps in the data with daily and weekly cycle}
\usage{
fill_large_gaps(trailing_dat, variable_name, daily_only = FALSE)
}
\arguments{
\item{trailing_dat}{data.frame with trailing NAs (to fill)}

\item{variable_name}{name of the column to fill}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
fills gaps in the data with daily and weekly cycle
}
