% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{fill_small_gaps}
\alias{fill_small_gaps}
\title{Fill small (length up to length 2) gaps in the data by linear interpolation}
\usage{
fill_small_gaps(ts_dat, variable_name)
}
\arguments{
\item{ts_dat}{the data.frame fo fill}

\item{variable_name}{variable to fill}
}
\description{
Fill small (length up to length 2) gaps in the data by linear interpolation
}
