% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{gap_fill_locations}
\alias{gap_fill_locations}
\title{gap fill a list of locations}
\usage{
gap_fill_locations(dat_list, variables, fill_until, daily_only = FALSE)
}
\arguments{
\item{dat_list}{list of data.frames,
each with one location to gap fill}

\item{variables}{list of variables to fill}

\item{fill_until}{time point to stop filling, if NULL the max of ts_dat is used.}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
gap fill a list of locations
}
