% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_gap_filll_trailing_NAs.R
\name{gap_fill_trailing_na_list}
\alias{gap_fill_trailing_na_list}
\title{fill gaps in a loop over an ordered list, each element with trailing NAs at the end
The functions starts at the beginning of the list and fills the gaps forward in sequence
thus the list has to be ordered by time}
\usage{
gap_fill_trailing_na_list(ts_dat_list, variable_name, daily_only = FALSE)
}
\arguments{
\item{ts_dat_list}{ordered list of trailing NA blocks}

\item{variable_name}{variable_name for gap filling}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
fill gaps in a loop over an ordered list, each element with trailing NAs at the end
The functions starts at the beginning of the list and fills the gaps forward in sequence
thus the list has to be ordered by time
}
