% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{gap_fill_variable}
\alias{gap_fill_variable}
\title{gap fill the data of one variable}
\usage{
gap_fill_variable(variable_name, ts_dat, fill_until, daily_only = FALSE)
}
\arguments{
\item{variable_name}{the variable to fill}

\item{ts_dat}{data.frame of one time series to fill}

\item{fill_until}{time point to stop filling, if NULL the max of ts_dat is used.}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
gap fill the data of one variable
}
