% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{gap_fill_variables}
\alias{gap_fill_variables}
\title{gap fill the data of one variable}
\usage{
gap_fill_variables(ts_dat, variables, fill_until, daily_only = FALSE)
}
\arguments{
\item{ts_dat}{data.frame of one multiple time series to fill}

\item{variables}{list of variables to fill}

\item{fill_until}{time point to stop filling, if NULL the max of ts_dat is used.}

\item{daily_only}{consider only a daily cycle for gap filling, no weekly cycle}
}
\description{
gap fill the data of one variable
}
