% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03d_nulls_back_to_null.R
\name{last_three_gaps_to_NA}
\alias{last_three_gaps_to_NA}
\title{Set last three gaps back to NA}
\usage{
last_three_gaps_to_NA(ts_dat, variable_name)
}
\arguments{
\item{ts_dat}{data.frame with the variable, the _filled column of the variable
and the _was_null column of the variable (information if the variable was NULL
in the raw data)}

\item{variable_name}{str name of the variable to be corrected}
}
\value{
data.frame with same number of rows and columns, but the last three
gaps of the variable will be NA again (so they are ready to be filled with
the ARIMA method)
}
\description{
When a new gap starts, it's first a small gap (up to two missing values).
They are filled with the previous value.
But in may become a larger gap. Then we want it to be filled using ARIMA.
The filling values we computed before are already used and not recomputed per default.
Without this function here, larger gaps would be treated as small gaps bit by bit.
Therefore we need to set the last gap to missing again after filling small gaps.
The consequence: If they become larger than 2, they will be filled with ARIMA as they should.
To make sure this is still working if the gap filling was down for some hours,
we treat the last three gaps.
}
