% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_fill_constant_variables.R
\name{more_than_one_unique_value}
\alias{more_than_one_unique_value}
\title{Inspect a data.frame for unique values}
\usage{
more_than_one_unique_value(df)
}
\arguments{
\item{df}{the data.frame to test}
}
\value{
vector with one element per data.frame column;
TRUE if all values in the column are identical (excluding NAs),
FALSE if there are at least two different values
}
\description{
Inspect a data.frame for unique values
}
