% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_sources.R
\name{name_data_sources}
\alias{name_data_sources}
\title{Name elements of a list of data sources}
\usage{
name_data_sources(data_sources_list)
}
\arguments{
\item{data_sources_list}{(list) an unnamed list of single sources}
}
\value{
list of named single sources, names correspond to elements'
$name slot.
}
\description{
Use the $name of a single source to name its element in a
a list of data sources.
}
