% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{remove_leading_na}
\alias{remove_leading_na}
\title{remove leading NAs from a data.frame}
\usage{
remove_leading_na(ts_dat, variable_name)
}
\arguments{
\item{ts_dat}{the data.frame}

\item{variable_name}{variable that should be used for removing leading NAs}
}
\description{
remove leading NAs from a data.frame
}
