% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_restrict_ranges.R
\name{restrict_value_ranges}
\alias{restrict_value_ranges}
\title{Restrict ranges of filled values to a meaningful range}
\usage{
restrict_value_ranges(
  locations_filled,
  allowed_ranges = allowed_ranges_filled_vars()
)
}
\arguments{
\item{locations_filled}{data frame containing some columns whose values will be restricted}

\item{allowed_ranges}{list with allowed ranges. Per default, the internal function
allowed_ranges_filled_vars() creates the list. The argument is only present to hand over a
different list for testing. If a boundary value is NA, the values remain unchanged.}
}
\description{
e.g., >= 0 for all pollutants
}
