% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_sources.R, R/01_single_source.R
\name{data_sources}
\alias{data_sources}
\alias{single_source}
\title{Data sources to query from clickhouse}
\usage{
data_sources()

single_source(
  name,
  variables,
  location_id_vars,
  get_data = fairqDbtools::send_query,
  load_mode = "replace",
  load_table = name,
  send_data = fairqDbtools::send_data,
  daily_only = FALSE,
  fill_until = NULL,
  optimize = FALSE,
  ...
)
}
\arguments{
\item{name}{(character) a unique name used as class and to identify target
table on database.}

\item{variables}{list of variables to gap-fill}

\item{location_id_vars}{a single id or a list of variables that provide a unique spatial key
(i.e. within each date_time) separating different locations to split the data into time series}

\item{get_data}{(function)}

\item{load_mode}{(character) defaults to "replace"}

\item{load_table}{(character) the table in that we load data}

\item{send_data}{(function)}

\item{daily_only}{if TRUE, use only daily_arima for gap filling, default FALSE}

\item{fill_until}{time point to stop filling; if NULL (default), the max of ts_dat is used.}

\item{optimize}{Should the source table be optimized (i.e. duplicate dropping) beforehand? FALSE by default.}

\item{...}{arguments used within methods and available for queries}
}
\description{
List data sources. This list defines which data source to extract from clickhouse.
}
