% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gap_filling.R
\name{split_trailing_na}
\alias{split_trailing_na}
\title{splits the data frame into groups with trailing consecutive NAs.
The data is arranged according to the date_time column, leading NAs are removed.
i.e. each group has first some values, then some trailing NA values}
\usage{
split_trailing_na(ts_dat, variable_name)
}
\arguments{
\item{ts_dat}{the data.frame for splitting}

\item{variable_name}{variable to split}
}
\description{
splits the data frame into groups with trailing consecutive NAs.
The data is arranged according to the date_time column, leading NAs are removed.
i.e. each group has first some values, then some trailing NA values
}
