% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03d_nulls_back_to_null.R
\name{undelivered_values_back_to_na}
\alias{undelivered_values_back_to_na}
\title{Set values that were originally NULL back to NA}
\usage{
undelivered_values_back_to_na(locations_filled, variables)
}
\arguments{
\item{locations_filled}{data frame with the original columns, the _filled columns, and
the _was_null columns containing the information whether the original value
delivered by the data source was NULL (0/1)}

\item{variables}{variables that were filled}
}
\value{
data frame where the original columns are corrected such that values are NULL
again if they were not delivered; the "..._was_null" columns are dropped.
}
\description{
Some values are not delivered in the original data.
We want them to be NULL again in the columns without "_filled" suffix in the target table.
However, in the query we already replace missing input values with the filled value
if it exists, in order to avoid filling the same gaps again and again in every gap filling.
This function uses the information from the "..._was_null" columns to set the
original columns back to NA (which will be NULL in the database).
We still need this information for model evaluation: Here we want to have the choice
whether we compare the predicted values with all values (including those we filled ourselves)
or only with the values that were originally delivered by the data source.
}
