% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_gap_filll_trailing_NAs.R
\name{weekly_arma}
\alias{weekly_arma}
\title{fits an ar(1)ma(1) to a time_series with a daily and a weekly cycle component}
\usage{
weekly_arma(time_series, n_ahead)
}
\arguments{
\item{time_series}{time_series object created with ts()}

\item{n_ahead}{number of obs to holdout and to make the prediction}
}
\description{
fits an ar(1)ma(1) to a time_series with a daily and a weekly cycle component
}
